backend/requirements.txt
fastapi
python-dotenv
python-multipart
uvicorn[standard]
pillow
ultralytics
pymongo